Code to properly adjust Anesthetic Medication inventory levels after adding, deleting or wasting quantities from the Anesthetic Record; added field QtyOnHandOld to table anestheticmedsgiven
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@2191 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/TableTypes/AnesthMedsInventoryAdj.cs
@@ -8,7 +8,7 @@ namespace OpenDentBusiness{
 		public int AdjustNum;
 		///<summary>FK to anesthmedsinventory.AnestheticMedNum.</summary>
 		public int AnestheticMedNum;
-		public int QtyAdj;
+		public double QtyAdj;
 		///<summary>FK to provider.ProvNum.</summary>
 		public int UserNum;
 		public string Notes;
